refactor(test): split component tests by domain
1. Move component tests into component/ sub-package
2. Split single ComponentTest.kt into BasicTest, ReconnectTest,
   CallbackTest, LifecycleTest, MiscTest
3. Move MockTransport.kt to component/ sub-package
4. Follow client-go naming convention for domain-based test files
